feat(tizen-remote): adds a library to interact with remote control ws server
This is very much a first pass, but seemed like it would be the minimal API we'd need as based on how we use it in `appium-tizen-tv-driver`.

I haven't yet used this lib in the driver--that's next.

Tangentially, I wanted to see if [typedoc](https://typedoc.org) was any good.  I was able to set up automatic doc generation pretty quickly.  Try `npm run docs` (or `npx typedoc`) in `packages/tizen-remote`, and see the `doc` output folder.