test(tizen-remote): refactor to support live devices
This (mostly) completes the work to allow the E2E tests to work with a real device instead of a mock server.

The following env vars are needed:

- `TEST_TIZEN_REMOTE_HOST` e.g., 127.0.0.1
- `TEST_TIZEN_REMOTE_PORT` e.g. 8002
- `TEST_TIZEN_REMOTE_TOKEN` - if you have one

if there is no token set, the test will get one and print it via debug message.

Changed the default name to `Appium` to be in sync with `appium-tizen-tv-driver`