chore(tizen-remote): fix leaks, update API
This is technically breaking, but I changed the signature of the `TizenRemote` constructor so that the "options" object is a fully optional second arg, and the only required arg (`host`) is first.

The main aim here was to write a handful of unit tests, but they of course uncovered a few problems.