chore(tizen-sample-app): create sample app
This is a sample app for a Tizen TV device.  Tests will use this as a fixture.

The interesting stuff is in `index.html` and `js/main.js`.  There are some npm scripts here to aid deployment.

This package is currently not intended to be published, but that could change in the future.