fix(appium-tizen-tv-driver,tizen-remote): do not always force token
the only time when we should specifically force a new token is when `appium:resetRcToken` is `true`.

also, immediately after instantiation, unless a token has been provided to `TizenRemote` (or via the env), the `hasToken()` method may resolve `true`, but it will not actually _set_ the token in memory from the cache due to it being a naughty side-effect.  unless `getToken()` was explicitly called, the token would not be set in-memory from the cache before the connection attempt.  the logic has changed so that, at connection time, we attempt to load the token from cache if one exists (and cache persistence is enabled).  if one still doesn't exist after connection, then it will ask for a new one.

also fixes some typescript problems which may actually be a bug in typescript. https://github.com/microsoft/TypeScript/issues/50286

adds a (slow) e2e test suite for checking persistence works