fix(appium-tizen-tv-driver,tizen-sample-app): add delay post-click for RC clicks
There's some delay of _n_ milliseconds between when the websocket server receives a "press key" command and when the associated `KeyboardEvent` actually makes it into the app.  There are more complicated/robust ways to ensure the `KeyboardEvent` is delivered, but for now we may be able to get away with a subsecond delay post-click.

Clicking via Chromedriver introduces no extra delays.

Updated `tizen-sample-app` to make info retrieval a bit easier.