feat(tizen-remote): update token when new one received
Previously, the lib only updated the cached token (when persistence is enabled) when a) it wasn't present, or b) the `resetRcToken` cap was `true`.

Evidently (?) tokens can expire, so this adds a listener for _unexpected_ "new token" messages and updates the internal token and cache (if applicable) when such a msg is received.

- Also renamed `lib/index.js` to `lib/tizen-remote.js` for sanity purposes
- Expose getters `token` and `tokenCachePath`