feat(tizen-remote): use per-host token storage
BREAKING CHANGE: Uses a new scheme for token storage. Instead of a single JSON file, we use one file per device.  This will break existing token stores.  Also, the `name` option of the `TizenRemote` constructor is no longer recognized.

This replaces the bespoke token cache with the `@appium/strongbox` package.  As such, it mitigates the need for lockfiles.