During testing on windows we found times when the strings returned from applist contained extra spaces which would cause problems.  When this happened the method _parseListAppsCmd would return no results which would cause automation to fail because the expected app was not installed.
Validated locally with a device which was showing the issue.

Added an additional unit test which had a string both with added spaces as well as trash on one line.  This allowed us to verify if we just see spaces at the end we are ok, but if we see trash then the line will be ignored.