fix: improve app list parsing on windows
Validated locally with a device which was showing the issue.

Added an additional unit test which had a string both with added spaces as well as trash on one line.  This allowed us to verify if we just see spaces at the end we are ok, but if we see trash then the line will be ignored.